requirements.txt
fastapi
uvicorn
requests
pypdf2
replicate
pymupdf
python-dotenv
azure-storage-blob

